Split into ClojureBase and Clojure plugins
The core JVM language plugins each have a base and a non-base plugin.
Base provides default values for source sets, compile tasks, etc. The
non-base plugin provides an opinionated setup, so provide a couple
source sets, add a test task, or documetnation task.

This split in the Clojure one, also relies on the JavaBase and Java
plugins. Gradle had mentioned "reacting" to the presence of these
plugins, which presumably means using the plugins.withId guard to only
apply certain logic if Java is applied. However, this is unlikely to
work out to great in our case, since many of the Java conventions are
ones that apply to Clojure as well. We may need to follow up on that
with them.

Additionally, this involved converting some of our plugin code and
source set code to align more with the out of the box plugins (Groovy,
being the most pertinent example).

This resulted in also separating the Clojure compilation output into
it's own classes dir. This should mean we can clean up some of the code
in the ClojureCOmpile task that created a temp destination dir to avoid
conflicts with Java code that could be compiled in the same destination.

This resolves #11 and #14.