Use the worker API for compilation
Compilation is now submitted via the worker API. The main issues right
now are that we are including a Clojure runtime on the classpath of the
worker in addition to the one on your compile classpath. A little test
seemed to show that it was still loading the right one (the one from the
compile classpath).

We could also try this for running tests.

Ideally, we could run without workerExecutor.await() but for some reason
it doesn't seem to make the actions wait like the docs say.

This is part of #9.