Detect clojure runtime to avoid version conflicts
For the ClojureCompile use of worker API, we should make sure you end up
with a different worker for a different version of Clojure. This change
detects the Clojure JAR on the compile classpath and uses that version
on the worker classpath.

There still seems to be a state issue that's causing the ClojureTest to
not work great with workers yet. So holding off on changing that.