Better isolation of Clojure runtimes in workers
Workers gain efficiency through reuse of the warm JVM (in addition to
potential parallelization). For Clojure, the reuse saves a bunch of
time, as the Clojure runtime is costly to start. In some contexts, the
runtime can get messy and carry state that can impact a later worker
task. For example, a Java class depended on by Clojure code will get
loaded into the runtime's classloader. If that class is later deleted,
and the worker is reused, the class is still there, meaning you won't
get a NoClassDefFoundError like you would expect.

By leveraging shimdandy to isolate the runtimes, we have more
flexibility to reuse the JVM, but potentially restart the Clojure
runtime. By additionally using system properties to communicate some
more information to the JVM, Gradle's worker reuse algorithm will limit
the situations we reuse the JVM.

As part of this, we also more explicitly leverage the tools JAR by
resolving it at execution-time of the build and only adding it to the
Clojure runtime classpath in the workers, not putting it in Gradle's
classloader. We can also more explicitly leverage prebuilt Clojure
scripts for task work, rather than building the script up through String
concatenation.

This is also fixes the reflection behavior to make the tests pass.