Publish to Bintray
In order for the tools JAR approach to work, we'll need it uploaded to a
repository someone would be willing to put in their project's
repositories block. JCenter being the easiest for us to take care of
(though potentially, Clojars as an option too). The group ID needs to
change for JCenter inclusion. They require a 3-segment inclusion rule
from a Bintray package. Since we have two JARs that both need to be
included, we would need the group ID to be 3-segment.