Add sources to dev runtime classpath
In order for REPL workflows to be helpful, you need to reload changes to
a namespace right after you edited it. This means we need the source
directories on the classpath in order for this to work.

Note that because of how our worker processes use shims, people using
tools.namespace as part of their workflow need to be on 0.3.0. This
switched to the use of java.classpath to identify files to refresh.