Removing source set outputs from classpath
With the outputs on the classpath, you end up with multiple versions of
the Clojure source and/or also AOT classes sitting there.

This opens up a "bug" where if you had Java source in the main source
set, you wouldn't have those compiled classes available in the REPL.
While that's not the intended behavior, this makes the REPL more useful
initial use cases. We'll have to fix this later.