Break up the basic clojure test
The Gradle TestKit seems to hold on to worker processes longer than it
should (well, at least for our case). This causes memory issues in CI
environments running our compat test suite. It seems to happen on the
basic-clojure test NS only (or at least first).

Watching the behavior locally, it seemed like it was getting up to 3.5+
GB before dropping down, and it dropped down after each Gradle version
compatTest task was over.

So it seemed to be tied to the test worker JVM being cleaned up. So step
1 was change the compat tests to fork on every test class for the compat
tests in a CI environment. Step 2 was to break up the basic-clojure test
NS to be smaller.

We'll see if it helps.