Fix block style configuration for ClojureCompile
You weren't able to do something like this:

  clojureCompile {
    options {
      reflectionsWarnings {
        enabled = true
      }
    }
  }

This happened both because we didn't provide an options(Action) method
on ClojureCompile for the top level one (you always had to do
options.reflectionWarnings), but also because, while we had a
reflectionWarnings(Action), Gradle does not instrument those with
Closure compatible methods like it does the top level methods in the
Task. We had to introduce our own overloads for this.

This fixes #63.