Try reusing workers more (again)
We used to force workers to be unique by Clojure/shim classpath, this
was mostly for trying to reuse the Clojure runtimes themselves. We're
not trying that at the moment, so it seems like we could maybe get some
improved performance by spinning up fewer worker JVMs.

By removing the shim classpath from system properties, that should mean
that there isn't anything project specific anymore about the workers
properties. This *may* mean that in our TestKit tests, the workers could
be reused across test projects. We'll have to try that out. But even so,
it did seem to make the compat tests run a little faster locally.
Anecdotal, though.