Sparser Gradle compatibility testing range
Instead of testing all versions between our min and max supported Gradle
versions, we'll just test the min and max (along with the next rc, when
available). Ideally we'd go back to testing them all, but the test suite
takes 45 minutes on Circle CI right now, which is just insane.