Ensure all main/test dirs are watched by Figwheel
Prior to this change, only the dev clojurescript source was added to the
watch dirs. This change puts all of the CLJ and CLJS in the dev source
set so it all shows up in the watch dirs.

Figwheel helper ns now also ensures that it only puts existing
directories in there. This happens within the REPL, so you can still
create the directory after the REPL starts.