Only re-root figwheel opts when key is present
Previously we were calling update-in on a bunch of settings, even if
they weren't already present, which caused them to be present with
null/empty values OR failing due to NPEs.

Now we check that the key exists before calling update-in.

A big use case for fixing this is proper support of modules with
Figwheel, since you end up without an output-to in the root compiler
options.

This fixes #104.