Fix inclusion of Clojure source in output when no AOT
When not AOTing, we want the Clojure source to be considered the output
of the source set. We had two issues in here.

1. The logic in the base plugins for CLJ and CLJS was backwards. It was
including the source only if you configured the compiler, rather than
the other way around.
2. The isCompilerConfigured for ClojureBuild was wrong. It treated an
empty set as a configured compiler. In Gradle 5, it changed the initial
state of SetProperty to be null instead of empty, highlighting this
issue.