Removing build settings for CI
We had a bunch of settings to deal with memory pressure issues in Circle
CI. Everything builds fine with them removed, so we'll try that for a
while. If we need to add these back, we'll have to change the location
of the stutter temp dir, since Gradle now blows up if you have a project
nested under another one (even in the build dir) without a
settings.gradle.