Add script to update wrapper in templates
The Gradle wrapper is an easy thing to forget as part of updating the
plugin. You may catch it when updating the build itself, but the
templates (unfortunately) have the wrapper manually copied into their
src dirs so that they are packaged into the JARs the templates get
distributed within.

An attempt was made to copy as part of the Gradle build, but it ran into
a locking issue, where presumably during the build Gradle is locking
some of the wrapper files.

This makes it easy to just run the script to update the version of
Gradle packaged in the templates to whaterver version the greater
Gradle-Clojure build uses.

The build will no also validate that the templates are using the same
version as the overall build to avoid them getting out of sync.