Rename plugin to clojurephant
Having a GitHub org starting with gradle caused some problems with
publishing to the plugin due to an irritating change that started
enforcing a namespace reserved for Gradle, which was anything containing
github.com/gradle. So even if your org name was gradle-clojure, you know
just as an example, and it's been common practice for years for plugins
to be named gradle-something, this was now prevented.

After trying to get a change to this in the forum [1], I was sick of
waiting for a fix given the month-plus wait time since reporting the
issue.

clojurephant is a reference to the Gradlephant logo.

[1] https://discuss.gradle.org/t/plugin-publish-fails/31598/13