Fix deprecation of deps on runtimeClasspath
Moving the dependency constaints to the devImplementation and devRuntimeOnly configurations to avoid a deprecation that will break in Gradle 7.

Since you can no longer directly declare (or it seems add constraints) to the runtimeClasspath config for a source set, we need to do this on each config that comprises it.

This fixes #136.