Use the out-of-box nrepl main
Switching to the nREPL main namespace provides a lot of small benefits,
mostly improving consistency with other nREPL environments. This also
removes some custom code and an extra port we had open to send the stop
message to the REPL.

This will require some modifications to Figwheel integration to stop
relying on the context map that was possible to send into the nREPL.

This also results in the .nrepl-port file being output, fixing #138.