Rework AOT logic to use a separate variant
The java-library plugin ends up triggering the consumption of a specific
variant of a dependency project. In particular, it grabs the classes
secondary variant from the apiElements configuration. Since the source
dir wasn't registered as a classes dir, it wouldn't show up in that
variant and then wouldn't be on the classpath of a downstream project.

Introducing a new variant for aotElements, simplifies the logic in the
plugin, so that it doesn't need to make decisions about which directory
to use anymore.

This fixes #128 and fixes #134.