Fix for multi-project classpath issues
The java-library plugin causes selection of a secondary classes variant
of the project dependency, which will cause only the registered
classesDirs to be put on the classpath.

In order to make this work the simplest, we need to have the Clojure
plugins register their source directories as classes dirs, if the
compiler isn't configured.

This fixes #128 and #134.