patch(common): Handle AsynchronousCloseException in prepl
It seems I'm getting a wider exception in the Prepl input loop
sometimes, this AsynchronousCloseException instead of the
ClosedByInterrupteException (which is a subclass) we used to get and
were handling. This just caused some harmless, but confusing, stack
traces in the Gradle output. This change should help clean that up.