major(common): Simplify repl classpath
This removes special logic we had to disable Clojure tasks when the repl
was the only task requested at the command line. Instead, we're just
being more explicit about which stuff we put on the classpath, and avoid
having dependencies on any of the *classes tasks that pulled in all
language compile tasks.

The breaking change here is that compiled test Java classes will not be
on the classpath anymore, however that doesn't seem like a common need.

It will also be a little more fiddly to get other language classes on
the classpath, but won't be terrible.

This does revert b151ff15ce7689224a3be3eca97c96e89c7c92e2, so it will
require some other solution to allow you to get your ClojureScript
compiles on the classpath.

I think this involves outputting to build/tmp/clojureRepl, but need to
make that easier and let people get access to all of their build
configurations.