minor(cider): Support jack-in of dependencies
Editor integrations, such as Cider and Calva (maybe others), rely on
being able to inject dependencies when starting nREPL in order to ensure
they have all of the capabilities their editors use.

Eventually, this could be supported natively by Gradle via [1], but this
has been a long-nagging issue with Gradle that made for a poorer user
experience than Leiningen, Boot, tools.deps, etc.

Since this is just a Gradle property, it can be provided by editors even
if the project isn't using Clojurephant (or the right version of
Clojurephant), meaning a graceful degradation to the old experience of
requiring the user to have put the right deps on the nREPL classpath.

[1] https://github.com/gradle/gradle/issues/17251

This fixes Clojurephant's side of #78.