minor(repl): Auto-download sources/javadoc for nREPL classpath
Cider has features that rely on the presence of javadoc and sources JARs
on the classpath. For Lein, that requires enrich-classpath [1], but we
can do it pretty straightforwardly in Gradle.

There wasn't a natural place to expose a toggle to disable this, and I
didn't want to introduce a new extension just for this. If someone asks,
we can do that though.

[1] https://github.com/clojure-emacs/enrich-classpath