patch(repl): Ensure cljs source on REPL classpath
There was a mistake in the new logic for putting source on the REPL
classpath, that meant a project using both Clojure and ClojureScript
plugins would only end up with one on the classpath.

This new behavior delays evaluation of which source to include and is
now safe for both plugins to be applied.