Remove support for configuration cache
Gradle's configuration cache feature is still incomplete and a moving
target. The maintenance burden of trying to stay in sync with it is
not worth the minor performance benefit it provides.

If they ever fully stabilize this feature, we can try adding support
back again.

I considered just removing the use of the removed internal API that we
previously used to communicate that the file collection was observed
to compute our inputs, but it seems to result in other class cast
issues that I don't want to investigate.

Fixes #206