Handle race in binding prepl promise
We use a promise to have the prepl now when a connection was made from
the client. Since this happens concurrently, @shilder has reported
that this can result in the connected var is created but not bound.

We already had a loop in place to retry in cases where the client
connected too early, but not for this scenario where the client gets
in after the var exists but before its bound to the promise.

From the sample error provided we receive a ClassNotFoundException, so
adding that to the catch block should cover the scenario.

Fixes #208 (hopefully)